backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic[email]==2.3.0
sqlalchemy[asyncio]==2.0.21
asyncpg==0.28.0
redis==5.0.0
paho-mqtt==1.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.24.1
pandas==2.1.0
numpy==1.25.2
psycopg2-binary==2.9.7
